argus_pipe_filter 最新の反映. 改行コード変更含む

+++ argus_synchro_lib/python/src/visualizer_module.cpp
@@ -61,7 +61,7 @@ void bind_visualizer(py::module& m)
         .def("summary",
              [](GodotUIVisualizer& visualizer, int isslow, const Eigen::Ref<const Eigen::MatrixXd>& boxes,
                 const Eigen::Ref<const Eigen::MatrixXf>& minmax, const Eigen::Ref<const Eigen::VectorXi>& valid_detects,
-                OctoTree octotree_obj, int angle_deg, const std::vector<py::array_t<uint8_t>>& frames,
+                     OctoTree octotree_obj, double angle_deg, const std::vector<py::array_t<uint8_t>>& frames,
                 const std::vector<CameraDetectionData>& bb_box_data, const std::vector<Camera>& camera,
                 Ccol_res collision_clusters, const std::map<int, NodeEntity>& cluster2entity, int ref_t,
                 int max_tree_depth, int dialate_point_size, bool octotree_func_on, bool collisiondetection_func_on,
